gui: save memory in Segment
Reorder fields to remove padding
Replace net ID with cost
Shrink cost to int8_t

+++ src/grt/include/grt/GlobalRouter.h
@@ -333,7 +333,7 @@ class GlobalRouter
   std::vector<Pin*> getAllPorts();
   void computeTrackConsumption(const Net* net,
                                int& track_consumption,
-                               std::vector<int>*& edge_costs_per_layer);
+                               std::vector<int8_t>*& edge_costs_per_layer);
 
   // aux functions
   std::vector<odb::Point> findOnGridPositions(const Pin& pin,
